refactor Math xforms out of MathDescription, and deduplicate code
