+++ vcell-core/src/main/java/cbit/vcell/math/MathDescription.java
@@ -1070,344 +1070,7 @@ private MathCompareResults compareInvariantAttributes(MathDescription newMathDes
 }
 
 
-/**
- * Insert the method's description here.
- * Creation date: (10/9/2002 10:54:06 PM)
- * @return cbit.vcell.math.MathDescription
- */
-public static MathDescription createCanonicalMathDescription(MathSymbolTableFactory mathSymbolTableFactory, MathDescription originalMathDescription) throws MathException, ExpressionException {
-	//
-	// clone current mathdescription
-	//
-	MathDescription newMath = new MathDescription(originalMathDescription);
-
-	newMath.makeCanonical(mathSymbolTableFactory);
-
-	return newMath;
-}
-
-
-/**
- * Insert the method's description here.
- * Creation date: (10/9/2002 10:54:06 PM)
- * @return cbit.vcell.math.MathDescription
- */
-public static MathDescription createMathWithExpandedEquations(MathDescription originalMathDescription, HashSet<String> varNamesToKeep) throws MathException, ExpressionException {
-	//
-	// clone current mathdescription
-	//
-	MathDescription newMath = new MathDescription(originalMathDescription);
-
-	//
-	// for any dependent variables in the "varNamesToKeep" list, create appropriate Variable/Equation
-	//
-	// this assumes that the dependent 'Function' is a linear combination of existing state variables
-	//  e.g. Function depVar = K0 + K1*indepVar1 + K2*indepVar2 + ... + Kn*indepVarN
-	// if it doesn't fit this form, then math's are not equivalent.
-	//
-	HashSet<String> stateVarSet = newMath.getStateVariableNames();
-	// Build the list of variables to be added
-	ArrayList<Function> varsToAdd = new ArrayList<Function>();
-	for (Variable var : newMath.variableList){
-		if (varNamesToKeep.contains(var.getName()) && var instanceof Function){
-			varsToAdd.add((Function)var);
-		}
-	}
-	for (Function function : varsToAdd){
-			//
-			// get list of symbols that are state variables
-			//
-			ArrayList<Variable> indepVarList = new ArrayList<Variable>();         // holds the "indepVar's"
-			ArrayList<Expression> coefficientList = new ArrayList<Expression>();      // holds the "K's"
-			Expression exp = function.getExpression();
-			exp.bindExpression(null);
-			Expression K0 = new Expression(exp);
-			K0.bindExpression(null);
-			String symbols[] = exp.getSymbols();
-			for (int j = 0; j < symbols.length; j++){
-				if (stateVarSet.contains(symbols[j])){
-					//
-					// store the independent variable (indepVar_i)
-					//
-					indepVarList.add(newMath.getVariable(symbols[j]));
-					Expression differential = exp.differentiate(symbols[j]);
-					differential = differential.flatten();
-					//
-					// store the coefficient (K_i)
-					//
-					coefficientList.add(differential);
-					//
-					// remove this term from the "constant" term (K0)
-					// (e.g. for expression "K0 + K1*V1 + ... + Ki*Vi + ... + Kn*Vn", Vi set to 0.0 and flattened)
-					// after each term is removed, only K0 is left
-					//
-					K0.substituteInPlace(new Expression(symbols[j]),new Expression(0.0));
-					K0 = K0.flatten();
-				}
-			}
-			//
-			// either all independent vars should be Volume, all should be Membrane, or all should be Filament 
-			//
-			int countVolumeVars = 0;
-			int countMembraneVars = 0;
-			int countFilamentVars = 0;
-			int countPointVars = 0;
-			for (int j = 0; j < indepVarList.size(); j++){
-				if (indepVarList.get(j) instanceof VolVariable){
-					countVolumeVars++;
-				}else if (indepVarList.get(j) instanceof MemVariable){
-					countMembraneVars++;
-				}else if (indepVarList.get(j) instanceof FilamentVariable){
-					countFilamentVars++;
-				}else if (indepVarList.get(j) instanceof PointVariable){
-					countPointVars++;
-				}else{
-					throw new RuntimeException("create canonicalMath cannot handle dependent vars of type '"+indepVarList.get(j).getClass().getName()+"'");
-				}
-			}
-			SubDomain functionSubdomain = (function.getDomain()!=null) ? newMath.getSubDomain(function.getDomain().getName()) : null;
-
-			//
-			// case: Volume Variable
-			// create VolVariable
-			// for each CompartmentSubDomain, create OdeEquation
-			// remove Function
-			//
-			if ((countVolumeVars > 0 && countVolumeVars == indepVarList.size()) || functionSubdomain instanceof CompartmentSubDomain){
-				VolVariable volVariable = new VolVariable(function.getName(),function.getDomain());
-				newMath.variableList.remove(function);
-				newMath.variableList.add(volVariable);
-				newMath.variableHashTable.remove(function.getName());
-				newMath.variableHashTable.put(volVariable.getName(), volVariable);
-				//
-				// determine which volume subdomains (CompartmentSubDomains) to add equations to
-				//    if domain information is available, then restrict the creation of equations to that subdomain,
-				//    else add equation to all subdomains of the same type (here: any CompartmentSubDomain)
-				//
-				Set<CompartmentSubDomain> compartmentSubDomains = new LinkedHashSet<>();
-				if (functionSubdomain != null){
-					compartmentSubDomains.add((CompartmentSubDomain) functionSubdomain);
-				}else if (countMembraneVars > 0){
-					for (Variable indepVar : indepVarList){
-						if (indepVar instanceof VolVariable && indepVar.getDomain() != null){
-							compartmentSubDomains.add((CompartmentSubDomain) newMath.getSubDomain(indepVar.getDomain().getName()));
-						}
-					}
-				}else{
-					for (SubDomain subDomain : newMath.getSubDomainCollection()){
-						if (subDomain instanceof CompartmentSubDomain){
-							compartmentSubDomains.add((CompartmentSubDomain) subDomain);
-						}
-					}
-				}
-				for (CompartmentSubDomain compartmentSubDomain : compartmentSubDomains){
-					//
-					// add an ODE where
-					//    initial value = K0 + Sum(coefficient_i*Var_i.init)
-					//    rate value = Sum(coefficient_i*Var_i.rate)
-					//
-					Expression initExp = new Expression(K0);
-					Expression rateExp = new Expression(0.0);
-					for (int k = 0; k < indepVarList.size(); k++){
-						Variable indepVar = indepVarList.get(k);
-						Equation indepVarEqu = compartmentSubDomain.getEquation(indepVar);
-						Expression coefficient = coefficientList.get(k);
-						initExp = Expression.add(initExp,Expression.mult(new Expression(coefficient),new Expression(indepVarEqu.getInitialExpression())));
-						rateExp = Expression.add(rateExp,Expression.mult(new Expression(coefficient),new Expression(indepVarEqu.getRateExpression())));
-					}
-					Variable initVariable = null;
-					for (int k=0;k<newMath.variableList.size();k++){
-						if (newMath.variableList.get(k).getName().startsWith(function.getName()+MATH_FUNC_INIT_SUFFIX_PREFIX)){
-							initVariable = newMath.variableList.get(k);
-							initExp = new Expression(initVariable,null);
-							break;
-						}
-					}
-					OdeEquation odeEquation = new OdeEquation(volVariable, initExp.flatten(), rateExp.flatten());
-					compartmentSubDomain.addEquation(odeEquation);
-				}
-			//
-			// case: Membrane Variable
-			//
-			}else if ((countMembraneVars > 0 && countMembraneVars == indepVarList.size()) || functionSubdomain instanceof MembraneSubDomain){
-				MemVariable memVariable = new MemVariable(function.getName(),function.getDomain());
-				newMath.variableList.remove(function);
-				newMath.variableList.add(memVariable);
-				newMath.variableHashTable.remove(function.getName());
-				newMath.variableHashTable.put(memVariable.getName(), memVariable);
-				//
-				// determine which membrane subdomains to add equations to
-				//    if domain information is available, then restrict the creation of equations to that subdomain,
-				//    else add equation to all subdomains of the same type (here: any MembraneSubDomain)
-				//
-				Set<MembraneSubDomain> membraneSubDomains = new LinkedHashSet<>();
-				if (functionSubdomain != null){
-					membraneSubDomains.add((MembraneSubDomain) functionSubdomain);
-				}else if (countMembraneVars > 0){
-					for (Variable indepVar : indepVarList){
-						if (indepVar instanceof MemVariable && indepVar.getDomain() != null){
-							membraneSubDomains.add((MembraneSubDomain) newMath.getSubDomain(indepVar.getDomain().getName()));
-						}
-					}
-				}else{
-					for (SubDomain subDomain : newMath.getSubDomainCollection()){
-						if (subDomain instanceof MembraneSubDomain){
-							membraneSubDomains.add((MembraneSubDomain) subDomain);
-						}
-					}
-				}
-				for (MembraneSubDomain membraneSubDomain : membraneSubDomains){
-					//
-					// add an ODE where
-					//    initial value = K0 + Sum(coefficient_i*Var_i.init)
-					//    rate value = Sum(coefficient_i*Var_i.rate)
-					//
-					Expression initExp = new Expression(K0);
-					Expression rateExp = new Expression(0.0);
-					for (int k = 0; k < indepVarList.size(); k++){
-						Variable indepVar = indepVarList.get(k);
-						Equation indepVarEqu = membraneSubDomain.getEquation(indepVar);
-						Expression coefficient = coefficientList.get(k);
-						initExp = Expression.add(initExp,Expression.mult(new Expression(coefficient),new Expression(indepVarEqu.getInitialExpression())));
-						rateExp = Expression.add(rateExp,Expression.mult(new Expression(coefficient),new Expression(indepVarEqu.getRateExpression())));
-					}
-					Variable initVariable = null;
-					for (int k=0;k<newMath.variableList.size();k++){
-						if (newMath.variableList.get(k).getName().startsWith(function.getName()+MATH_FUNC_INIT_SUFFIX_PREFIX)){
-							initVariable = newMath.variableList.get(k);
-							initExp = new Expression(initVariable,null);
-							break;
-						}
-					}
-					OdeEquation odeEquation = new OdeEquation(memVariable, initExp.flatten(), rateExp.flatten());
-					membraneSubDomain.addEquation(odeEquation);
-				}
-			//
-			// case: Filament Variable
-			//
-			}else if ((countFilamentVars > 0 && countFilamentVars == indepVarList.size()) || functionSubdomain instanceof FilamentSubDomain){
-				FilamentVariable filamentVariable = new FilamentVariable(function.getName(),function.getDomain());
-				newMath.variableList.remove(function);
-				newMath.variableList.add(filamentVariable);
-				newMath.variableHashTable.remove(function.getName());
-				newMath.variableHashTable.put(filamentVariable.getName(), filamentVariable);
-				//
-				// determine which filament subdomains to add equations to
-				//    if domain information is available, then restrict the creation of equations to that subdomain,
-				//    else add equation to all subdomains of the same type (here: any FilamentSubDomain)
-				//
-				Set<FilamentSubDomain> filamentSubDomains = new LinkedHashSet<>();
-				if (functionSubdomain != null){
-					filamentSubDomains.add((FilamentSubDomain) functionSubdomain);
-				}else if (countFilamentVars > 0){
-					for (Variable indepVar : indepVarList){
-						if (indepVar instanceof FilamentVariable && indepVar.getDomain() != null){
-							filamentSubDomains.add((FilamentSubDomain) newMath.getSubDomain(indepVar.getDomain().getName()));
-						}
-					}
-				}else{
-					for (SubDomain subDomain : newMath.getSubDomainCollection()){
-						if (subDomain instanceof FilamentSubDomain){
-							filamentSubDomains.add((FilamentSubDomain) subDomain);
-						}
-					}
-				}
-				for (FilamentSubDomain filamentSubDomain : filamentSubDomains){
-					//
-					// add an ODE where
-					//    initial value = K0 + Sum(coefficient_i*Var_i.init)
-					//    rate value = Sum(coefficient_i*Var_i.rate)
-					//
-					Expression initExp = new Expression(K0);
-					Expression rateExp = new Expression(0.0);
-					for (int k = 0; k < indepVarList.size(); k++){
-						Variable indepVar = indepVarList.get(k);
-						Equation indepVarEqu = filamentSubDomain.getEquation(indepVar);
-						Expression coefficient = coefficientList.get(k);
-						initExp = Expression.add(initExp,Expression.mult(new Expression(coefficient),new Expression(indepVarEqu.getInitialExpression())));
-						rateExp = Expression.add(rateExp,Expression.mult(new Expression(coefficient),new Expression(indepVarEqu.getRateExpression())));
-					}
-					Variable initVariable = null;
-					for (int k=0;k<newMath.variableList.size();k++){
-						if (newMath.variableList.get(k).getName().startsWith(function.getName()+MATH_FUNC_INIT_SUFFIX_PREFIX)){
-							initVariable = newMath.variableList.get(k);
-							initExp = new Expression(initVariable,null);
-							break;
-						}
-					}
-					OdeEquation odeEquation = new OdeEquation(filamentVariable, initExp.flatten(), rateExp.flatten());
-					filamentSubDomain.addEquation(odeEquation);
-				}
-				
-			//
-			// case: Point Variable
-			//
-			}else if ((countPointVars > 0 && countPointVars == indepVarList.size()) || functionSubdomain instanceof PointSubDomain){
-				PointVariable pointVariable = new PointVariable(function.getName(),function.getDomain());
-				newMath.variableList.remove(function);
-				newMath.variableList.add(pointVariable);
-				newMath.variableHashTable.remove(function.getName());
-				newMath.variableHashTable.put(pointVariable.getName(), pointVariable);
-				//
-				// determine which point subdomains to add equations to
-				//    if domain information is available, then restrict the creation of equations to that subdomain,
-				//    else add equation to all subdomains of the same type (here: any PointSubDomain)
-				//
-				Set<PointSubDomain> pointSubDomains = new LinkedHashSet<>();
-				if (functionSubdomain != null){
-					pointSubDomains.add((PointSubDomain) functionSubdomain);
-				}else if (countPointVars > 0){
-					for (Variable indepVar : indepVarList){
-						if (indepVar instanceof PointVariable && indepVar.getDomain() != null){
-							pointSubDomains.add((PointSubDomain) newMath.getSubDomain(indepVar.getDomain().getName()));
-						}
-					}
-				}else{
-					for (SubDomain subDomain : newMath.getSubDomainCollection()){
-						if (subDomain instanceof PointSubDomain){
-							pointSubDomains.add((PointSubDomain) subDomain);
-						}
-					}
-				}
-				for (PointSubDomain pointSubDomain : pointSubDomains){
-					//
-					// add an ODE where
-					//    initial value = K0 + Sum(coefficient_i*Var_i.init)
-					//    rate value = Sum(coefficient_i*Var_i.rate)
-					//
-					Expression initExp = new Expression(K0);
-					Expression rateExp = new Expression(0.0);
-					for (int k = 0; k < indepVarList.size(); k++){
-						Variable indepVar = indepVarList.get(k);
-						Equation indepVarEqu = pointSubDomain.getEquation(indepVar);
-						Expression coefficient = coefficientList.get(k);
-						initExp = Expression.add(initExp,Expression.mult(new Expression(coefficient),new Expression(indepVarEqu.getInitialExpression())));
-						rateExp = Expression.add(rateExp,Expression.mult(new Expression(coefficient),new Expression(indepVarEqu.getRateExpression())));
-					}
-					Variable initVariable = null;
-					for (int k=0;k<newMath.variableList.size();k++){
-						if (newMath.variableList.get(k).getName().startsWith(function.getName()+MATH_FUNC_INIT_SUFFIX_PREFIX)){
-							initVariable = newMath.variableList.get(k);
-							initExp = new Expression(initVariable,null);
-							break;
-						}
-					}
-					OdeEquation odeEquation = new OdeEquation(pointVariable, initExp.flatten(), rateExp.flatten());
-					pointSubDomain.addEquation(odeEquation);
-				}
-
-			//
-			//
-			//
-			}else{
-				throw new RuntimeException("create canonicalMath cannot handle mixture of dependent vars types");
-			}
-	}
-	
-	return newMath;
-}
-
-/**
+	/**
  * The firePropertyChange method was generated to support the propertyChange field.
  */
 public void firePropertyChange(java.lang.String propertyName, java.lang.Object oldValue, java.lang.Object newValue) {
@@ -1730,18 +1393,26 @@ protected java.beans.PropertyChangeSupport getPropertyChange() {
 	return propertyChange;
 }
 
-/**
- * Insert the method's description here.
- * Creation date: (11/24/2004 7:20:17 AM)
- * @return cbit.vcell.math.Variable[]
- */
-public HashSet<String> getStateVariableNames() {
-	HashSet<String> stateVarNameSet = new HashSet<String>();
+public List<Variable> getStateVariables() {
+	List<Variable> stateVars = new ArrayList<>();
 	for (int i = 0; i < variableList.size(); i++){
 		Variable var = variableList.get(i);
 		if (var instanceof VolVariable || var instanceof MemVariable || var instanceof FilamentVariable ||
-			var instanceof VolumeRegionVariable || var instanceof MembraneRegionVariable || var instanceof FilamentRegionVariable ||
-			var instanceof ParticleVariable || var instanceof StochVolVariable || var instanceof PointVariable){
+				var instanceof VolumeRegionVariable || var instanceof MembraneRegionVariable || var instanceof FilamentRegionVariable ||
+				var instanceof ParticleVariable || var instanceof StochVolVariable || var instanceof PointVariable){
+			stateVars.add(var);
+		}
+	}
+	return stateVars;
+}
+
+public HashSet<String> getStateVariableNames() {
+	List<Variable> stateVars = getStateVariables();
+	HashSet<String> stateVarNameSet = new HashSet<String>();
+	for (Variable var : stateVars){
+		if (var instanceof VolVariable || var instanceof MemVariable || var instanceof FilamentVariable ||
+				var instanceof VolumeRegionVariable || var instanceof MembraneRegionVariable || var instanceof FilamentRegionVariable ||
+				var instanceof ParticleVariable || var instanceof StochVolVariable || var instanceof PointVariable){
 			stateVarNameSet.add(var.getName());
 		}
 	}
@@ -1800,6 +1471,13 @@ public Variable getVariable(String name) {
 public Enumeration<Variable> getVariables() {
 	return Collections.enumeration(variableList);
 }
+
+	public List<Variable> getVariableList() {
+		return variableList;
+	}
+	public Map<String, Variable> getVariableMap() {
+		return variableHashTable;
+	}
 
 public Iterator<Event> getEvents() {
 	return eventList.iterator();
@@ -3750,64 +3428,9 @@ public static MathCompareResults testEquivalency(MathSymbolTableFactory mathSymb
 		if (mathDescription2.compareEqual(mathDescription1)){
 			return new MathCompareResults(Decision.MathEquivalent_NATIVE);
 		}else{
-			//
-			// must test for equivalence
-			//
-			HashSet<String> indepVars1 = mathDescription1.getStateVariableNames();
-			HashSet<String> indepVars2 = mathDescription2.getStateVariableNames();
-			HashSet<String> union = new HashSet<String>(indepVars1);
-			union.addAll(indepVars2);
-			
-//			MathDescription canonicalMath1 = MathDescription.createCanonicalMathDescription(mathSymbolTableFactory,createMathWithExpandedEquations(mathDescription1,union));
-//			MathDescription canonicalMath2 = MathDescription.createCanonicalMathDescription(mathSymbolTableFactory,createMathWithExpandedEquations(mathDescription2,union));
-
-			MathDescription canonicalMath1 = MathDescription.createMathWithExpandedEquations(mathDescription1,union);
-			MathDescription canonicalMath2 = MathDescription.createMathWithExpandedEquations(mathDescription2,union);
-			
-			
-			HashSet<String> canonIndepVars1 = canonicalMath1.getStateVariableNames();
-			HashSet<String> canonIndepVars2 = canonicalMath2.getStateVariableNames();
-			
-			for (String varName : union) {
-				if (!canonIndepVars1.contains(varName)) {
-					logger.error("========= indepVars1 size: "+indepVars1.size());
-					logger.error("========= union size: "+union.size());
-					logger.error("========= Canonical Math 1 is missing variable: "+varName);
-				}
-			}
-			for (String varName : union) {
-				if (!canonIndepVars2.contains(varName)) {
-					logger.error("========= indepVars2 size: "+indepVars2.size());
-					logger.error("========= union size: "+union.size());
-					logger.error("========= Canonical Math 2 is missing variable: "+varName);
-				}
-			}
-
-//			HashSet<String> depVarsToSubstitute1 = new HashSet<String>();
-//			for (Variable var : Collections.list(canonicalMath1.getVariables())) {
-//				depVarsToSubstitute1.add(var.getName());
-//			}
-//			HashSet<String> depVarsToSubstitute2 = new HashSet<String>();
-//			for (Variable var : Collections.list(canonicalMath1.getVariables())) {
-//				depVarsToSubstitute2.add(var.getName());
-//			}
-//			depVarsToSubstitute1.removeAll(union);
-//			depVarsToSubstitute2.removeAll(union);
-//			if (depVarsToSubstitute1.size()>0){
-//				System.out.println(depVarsToSubstitute1);
-//				String depVarNames[] = (String[])depVarsToSubstitute1.toArray(new String[depVarsToSubstitute1.size()]);
-//				Function functionsToSubstitute[] = MathDescription.getFlattenedFunctions(mathSymbolTableFactory,mathDescription1,depVarNames);
-//				System.out.println(functionsToSubstitute);
-//				canonicalMath1.substituteInPlace(mathSymbolTableFactory,functionsToSubstitute);
-//			}
-//			if (depVarsToSubstitute2.size()>0){
-//				String depVarNames[] = (String[])depVarsToSubstitute2.toArray(new String[depVarsToSubstitute2.size()]);
-//				Function functionsToSubstitute[] = MathDescription.getFlattenedFunctions(mathSymbolTableFactory,mathDescription2,depVarNames);
-//				canonicalMath1.substituteInPlace(mathSymbolTableFactory,functionsToSubstitute);
-//			}
-//			// flatten again
-			canonicalMath1.makeCanonical(mathSymbolTableFactory);
-			canonicalMath2.makeCanonical(mathSymbolTableFactory);
+			MathDescription[] canonicalMaths = MathUtilities.getCanonicalMathDescriptions(mathDescription1, mathDescription2);
+			MathDescription canonicalMath1 = canonicalMaths[0];
+			MathDescription canonicalMath2 = canonicalMaths[1];
 			// now compare
 			return canonicalMath2.compareEquivalentCanonicalMath(canonicalMath1);
 		}
